# lib/api-zod/src/generated/types/sectorKeywords.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Builder & ATS Optimizer API
 * OpenAPI spec version: 0.1.0
 */
import type { SectorKeywordsCategories } from "./sectorKeywordsCategories";

export interface SectorKeywords {
  sector: string;
  keywords: string[];
  categories: SectorKeywordsCategories;
}