# lib/api-zod/src/generated/types/updateResumeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Builder & ATS Optimizer API
 * OpenAPI spec version: 0.1.0
 */
import type { ResumeData } from "./resumeData";
import type { UpdateResumeBodyStatus } from "./updateResumeBodyStatus";

export interface UpdateResumeBody {
  title?: string;
  sector?: string;
  /** @nullable */
  templateId?: number | null;
  data?: ResumeData;
  status?: UpdateResumeBodyStatus;
}