# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Builder & ATS Optimizer API
 * OpenAPI spec version: 0.1.0
 */
import type { UserTier } from "./userTier";

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  sector?: string | null;
  onboardingCompleted: boolean;
  tier: UserTier;
  createdAt: Date;
}